Resolves #266
