+++ docs/app/(docs)/[[...slug]]/(home)/components/benefits/intuitive.tsx
@@ -3,7 +3,7 @@
 import { motion, useInView } from "motion/react";
 import { useRef } from "react";
 
-const command = "npx ultracite format";
+const command = "npx ultracite fix";
 const lines = [
   "Diagnostics not shown: 35.",
   "Checked 1740 files in 214ms.",
