+++ docs/content/migrate/eslint.mdx
@@ -110,8 +110,8 @@ If you have ESLint related scripts in your `package.json`, you can replace them 
 ```json title="package.json"
 {
   "scripts": {
-    "lint": "npx ultracite@latest lint",
-    "format": "npx ultracite@latest format"
+    "check": "npx ultracite@latest check",
+    "fix": "npx ultracite@latest fix"
   }
 }
 ```
