+++ docs/content/migrate/prettier.mdx
@@ -109,8 +109,8 @@ If you have Prettier related scripts in your `package.json`, you can replace the
 ```json title="package.json"
 {
   "scripts": {
-    "lint": "npx ultracite@latest lint",
-    "format": "npx ultracite@latest format"
+    "check": "npx ultracite@latest check",
+    "fix": "npx ultracite@latest fix"
   }
 }
 ```
