+++ docs/content/monorepos.mdx
@@ -27,8 +27,8 @@ You can add the `format` and `lint` scripts to your `package.json` file.
 
 {
   "scripts": {
-    "format": "npx ultracite@latest format",
-    "lint": "npx ultracite@latest lint"
+    "check": "npx ultracite@latest check",
+    "fix": "npx ultracite@latest fix"
   }
 }
 ```
