+++ scripts/commands/check.ts
@@ -2,16 +2,8 @@ import { spawnSync } from "node:child_process";
 import process from "node:process";
 import { parseFilePaths } from "../utils";
 
-type FormatOptions = {
-  unsafe?: boolean;
-};
-
-export const format = (files: string[], options: FormatOptions = {}) => {
-  const args = ["npx", "@biomejs/biome", "check", "--write"];
-
-  if (options.unsafe) {
-    args.push("--unsafe");
-  }
+export const check = (files: string[]) => {
+  const args = ["npx", "@biomejs/biome", "check"];
 
   // Add files or default to current directory
   if (files.length > 0) {
@@ -28,7 +20,6 @@ export const format = (files: string[], options: FormatOptions = {}) => {
   });
 
   if (result.error) {
-    // biome-ignore lint/suspicious/noConsole: "We want to log the error to the console"
     console.error("Failed to run Ultracite:", result.error.message);
     process.exit(1);
   }
